algorithm fixes

+++ src/main/kotlin/com/oast_projekt/model/Solution.kt
@@ -3,14 +3,12 @@ package com.oast_projekt.model
 data class Solution (
     val mapOfValues: MutableMap<Point, Int>
 ) {
-    var cost: Float? = null
+    var cost: Double? = null
     var capacitiesOfLinks: List<Int>? = null
     var numberOfLinksWithExceededCapacity: Int? = null
 
     val numberOfGenes: Int
-        get() = mapOfValues.entries
-            .map { entry -> entry.key.demandId }
-            .size
+        get() = mapOfValues.entries.map { entry -> entry.key.demandId }.size
 
     fun getGene(geneId: Int): Map<Point, Int> {
         val gene = mutableMapOf<Point, Int>()
